+++ ProjectAnnihilation/Assets/Scripts/UnitScripts/Projectile/ProjectileManager.cs
@@ -1,11 +1,9 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 public class ProjectileManager : MonoBehaviour
 {
     //Useful variables 
-    public int damageDone;
+    public float damageDone;
     public bool isAttacker;
     // Start is called before the first frame update
     void Start()
